src/requirements.txt
python-dotenv==1.1.1
python-docx==1.2.0
markitdown[docx]
pymupdf4llm==0.0.27
fastapi==0.116.1
uvicorn==0.35.0
retry==0.9.2
langchain-community==0.3.27
qdrant-client==1.15.1
pymongo==4.14.1
pandas==2.3.2
python-pptx==1.0.2
openai==1.101.0